Fix trailing whitespace
Not leaving trailing whitespace is good coding style and keeps
commits clean. A lot of editors (are set up to) remove trailing
whitespace automatically, and these lines will show up as
unrelated changes in the people's pull requests, unless they go
to the trouble of explicitly not comitting them.

Trailing whitespace is usually invisible and syntactically and
semantically irrelevant. This README.md is likely no exception.